move script cache to be an interface
